pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "sqlmodel>=0.0.27",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    # LangChain and LangGraph
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.1.11",
    # Streamlit
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
    "ipykernel>=7.0.1",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "faker>=24.0.0",
    "redis>=5.0.0",
    "aioredis>=2.0.0",
    "distutils-pytest>=0.2.1",
]
